+++ dwds/test/fixtures/fakes.dart
@@ -166,6 +166,10 @@ class FakeModules implements Modules {
   Future<Uri> libraryForSource(String serverPath) async => Uri(path: _library);
 
   @override
+  Future<Uri> libraryOrPartForSource(String serverPath) async =>
+      Uri(path: _library);
+
+  @override
   Future<String> moduleForSource(String serverPath) async => _module;
 
   @override
@@ -450,6 +454,7 @@ class FakeExpressionCompiler implements ExpressionCompiler {
   Future<ExpressionCompilationResult> compileExpressionToJs(
     String isolateId,
     String libraryUri,
+    String scriptUri,
     int line,
     int column,
     Map<String, String> jsModules,