+++ frontend_server_common/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.2.3-wip
 
-- Update minimum SDK constraint to 3.8.0.
+- Update Dart SDK constraint to `^3.10.0`.
 - Add bootstrapping code for DDC library bundle format.
+- Added scriptUri to compileExpression*Request
 
 ## 0.2.2